+++ src/Application/Adapters/Laravel/Commands/InsightsCommand.php
@@ -6,6 +6,7 @@ namespace NunoMaduro\PhpInsights\Application\Adapters\Laravel\Commands;
 
 use Illuminate\Console\Command;
 use NunoMaduro\PhpInsights\Application\Console\Commands\AnalyseCommand;
+use NunoMaduro\PhpInsights\Application\Console\Definitions\AnalyseDefinition;
 use NunoMaduro\PhpInsights\Domain\Kernel;
 use NunoMaduro\PhpInsights\Domain\Reflection;
 
@@ -17,7 +18,7 @@ final class InsightsCommand extends Command
     /**
      * {@inheritdoc}
      */
-    protected $signature = 'insights {directory?} {--config-path=config/insights.php}';
+    protected $name = 'insights';
 
     /**
      * {@inheritdoc}
@@ -40,8 +41,22 @@ final class InsightsCommand extends Command
 
         $output = (new Reflection($this->output))->get('output');
 
-        $analyseCommand->__invoke($this->input, $output);
+        return $analyseCommand->__invoke($this->input, $output);
+    }
 
-        return 0;
+    /**
+     * {@inheritdoc}
+     */
+    public function configure(): void
+    {
+        parent::configure();
+
+        $this->setDefinition(
+            AnalyseDefinition::get()
+        );
+
+        $this->getDefinition()
+            ->getOption('config-path')
+            ->setDefault('config/insights.php');
     }
 }