+++ tests/TestCase.php
@@ -13,14 +13,15 @@ abstract class TestCase extends BaseTestCase
     /**
      * Call protected/private method of a class.
      *
-     * @param string $class Instantiated object that we will run method on.
-     * @param string $methodName Method name to call.
-     * @param array $parameters Array of parameters to pass into method.
+     * @param  string  $class Instantiated object that we will run method on
+     * @param  string  $methodName Method name to call
+     * @param  array<int, mixed>  $parameters Array of parameters to pass into method
      *
-     * @return mixed Method result.
+     * @return mixed Method result
+     *
      * @throws ReflectionException
      */
-    public function invokeStaticMethod(string $class, string $methodName, ...$parameters)
+    public function invokeStaticMethod(string $class, string $methodName, array $parameters)
     {
         $reflection = new ReflectionClass($class);
         $method = $reflection->getMethod($methodName);